+++ TerrariaCells.csproj
@@ -16,14 +16,12 @@
 	  <None Remove="Common\Assets\MummySlam.png" />
 	  <None Remove="Content\Projectiles\MummyShockwave.png" />
 	</ItemGroup>
+  
 	<ItemGroup>
 	  <Reference Include="StructureHelper">
 	    <HintPath>..\ModAssemblies\StructureHelper_v2.5.2.dll</HintPath>
 	  </Reference>
 	</ItemGroup>
-	<PropertyGroup>
-		
-	</PropertyGroup>
 
 	<!-- References -->
 	<ItemGroup>
